+++ internal/controller/positions.go
@@ -24,6 +24,11 @@ func (c *PositionsController) GetPosition(w http.ResponseWriter, r *http.Request
 	}
 
 	positionID := r.PathValue("id")
+	if positionID == "" {
+		errorHandler(w, r, &HTTPError{Detail: "error at getting position: id is incorrect", Status: http.StatusBadRequest})
+		return
+	}
+
 	position, err := c.Repo.Get(r.Context(), positionID)
 
 	if err != nil {
@@ -85,6 +90,11 @@ func (c *PositionsController) DeletePosition(w http.ResponseWriter, r *http.Requ
 	}
 
 	positionID := r.PathValue("id")
+	if positionID == "" {
+		errorHandler(w, r, &HTTPError{Detail: "error at deleting position: id is incorrect", Status: http.StatusBadRequest})
+		return
+	}
+
 	err := c.Repo.Delete(r.Context(), positionID)
 
 	if err != nil {
@@ -103,7 +113,7 @@ func (c *PositionsController) UpdatePosition(w http.ResponseWriter, r *http.Requ
 
 	positionID := r.PathValue("id")
 	if positionID == "" {
-		errorHandler(w, r, &HTTPError{Detail: "missing position ID", Status: http.StatusBadRequest})
+		errorHandler(w, r, &HTTPError{Detail: "error at updating position: id is incorrect", Status: http.StatusBadRequest})
 		return
 	}
